feat: add working cholesky decomposition

+++ include/cpp/poisson.hpp
@@ -19,7 +19,7 @@ struct poisson_config {
   int Y_res = 6; // Number of points in y-direction
   double etol = 1e-6; // Error tolerance
   int max_iter = 5000; // Maximum number of iterations
-  Precision prec = Half; // precision for the solve
+  Precision prec = Single; // precision for the solve
   int num_experiments = 100; // number of experiments (number of times RHS is sampled)
   gamma_config gamma_cfg; // bounds config
   int block_jacobi_tile_size = 64; // block jacobi tile size
